Added Quality & Rarity

+++ database/migrations/2021_01_03_231322_create_component_ship_table.php
@@ -17,7 +17,7 @@ class CreateComponentShipTable extends Migration
             $table->foreignId('component_id');
             $table->foreignId('ship_id');
 
-            $table->integer('health');
+            $table->float('health');
         });
     }
